Reworked the run logic for modules - allow them to flag they won't run
If a module sets the self._donotrun flag to True, the module logic will
take care of not running this module at all.

This also sorts out some issues when hackbench is not run, where some of
the log file descriptors are not found when cleaning up in the end.

+++ rteval/modules/loads/hackbench.py
@@ -51,6 +51,8 @@ class Hackbench(CommandLineLoad):
         else:
             self._log(Log.INFO, "hackbench: low memory system (%f GB/core)! Not running\n" % ratio)
             mult = 0
+            self._donotrun = True
+
         self.jobs = self.num_cpus * mult
 
         self.args = ['hackbench',  '-P',
@@ -67,11 +69,6 @@ class Hackbench(CommandLineLoad):
 
 
     def _WorkloadPrepare(self):
-        # if we don't have any jobs just wait for the stop event and return
-        if self.jobs == 0:
-            self.WaitForCompletion()
-            return
-
         self.__nullfp = os.open("/dev/null", os.O_RDWR)
         if self._logging:
             self.__out = self.open_logfile("hackbench.stdout")
@@ -111,6 +108,9 @@ class Hackbench(CommandLineLoad):
 
 
     def _WorkloadCleanup(self):
+        if self._donotrun:
+            return
+
         if self.__hbproc.poll() == None:
             os.kill(self.__hbproc.pid, SIGKILL)
         self.__hbproc.wait()
